+++ src/helpers.ts
@@ -71,6 +71,8 @@ export const enrichSchema = (schema: BormSchema, dbHandles: DBHandles): Enriched
 							`[Schema] ${key} is extending a thing but missing the "as" property in its defaultDBConnector`,
 						);
 					}
+
+					/// IMPORT THE EXTENDED SCHEMA
 					const extendedSchema = draft.entities[value.extends] || draft.relations[value.extends];
 					/// find out all the thingTypes this thingType is extending
 					// @ts-expect-error allExtends does not belong to the nonEnriched schema so this ts error is expecte
@@ -80,6 +82,7 @@ export const enrichSchema = (schema: BormSchema, dbHandles: DBHandles): Enriched
 					value.idFields = extendedSchema.idFields
 						? (value.idFields || []).concat(extendedSchema.idFields)
 						: value.idFields;
+
 					value.dataFields = extendedSchema.dataFields
 						? (value.dataFields || []).concat(
 								extendedSchema.dataFields.map((df: DataField) => {
@@ -97,6 +100,7 @@ export const enrichSchema = (schema: BormSchema, dbHandles: DBHandles): Enriched
 								}),
 							)
 						: value.dataFields;
+
 					value.linkFields = extendedSchema.linkFields
 						? (value.linkFields || []).concat(extendedSchema.linkFields)
 						: value.linkFields;
@@ -112,6 +116,13 @@ export const enrichSchema = (schema: BormSchema, dbHandles: DBHandles): Enriched
 						if (Object.keys(val.roles).length === 0) {
 							val.roles = {};
 						}
+					}
+
+					//todo: Do some checks, and potentially simplify the hooks structure
+					if (extendedSchema?.hooks?.pre) {
+						value.hooks = value.hooks || {};
+						value.hooks.pre = value.hooks.pre || [];
+						value.hooks.pre = [...(extendedSchema?.hooks?.pre || []), ...(value?.hooks?.pre || [])];
 					}
 				}
 			},