+++ tests/unit/mutations/mutationPreHooks.test.ts
@@ -362,6 +362,44 @@ describe('Mutations: PreHooks', () => {
 		}
 	});
 
+	it('tn3[transform, inherited] Append children to node', async () => {
+		expect(bormClient).toBeDefined();
+
+		try {
+			const mutRes = await bormClient.mutate(
+				{
+					$thing: 'Kind',
+					id: 'secret-kind-tn3',
+					space: 'space-1',
+				},
+				{ noMetadata: true },
+			);
+
+			console.log('mutRes', mutRes);
+
+			const res = await bormClient.query(
+				{
+					$thing: 'Kind',
+					$thingType: 'relation',
+					$id: 'secret-kind-tn3-YES!',
+				},
+				{ noMetadata: true },
+			);
+			expect(res).toEqual({
+				id: 'secret-kind-tn3-YES!',
+				space: 'space-1',
+			});
+		} finally {
+			//clean
+			await bormClient.mutate({
+				$thing: 'Kind',
+				$thingType: 'relation',
+				$op: 'delete',
+				$ids: ['secret-kind-tn3', 'secret-kind-tn3-YES!'],
+			});
+		}
+	});
+
 	it('ctx1[transform, context] Use context', async () => {
 		expect(bormClient).toBeDefined();